--- src/features/ModelPlan/TaskList/Basics/Milestones/index.scss
@@ -1,13 +1,0 @@
-@use 'viewports' as *;
-
-.milestone-form {
-  .datepicker__wrapper {
-    display:grid;
-    grid-template-columns: 1fr 1fr;
-    column-gap: units(3);
-
-    @media screen and (max-width: $desktop) {
-      grid-template-columns: 1fr;
-    }
-  }
-}
