+++ src/gql/operations/Basics/GetAllBasics.ts
@@ -20,16 +20,6 @@ export default gql(/* GraphQL */ `
         goal
         testInterventions
         note
-        completeICIP
-        clearanceStarts
-        clearanceEnds
-        announced
-        applicationsStart
-        applicationsEnd
-        performancePeriodStarts
-        performancePeriodEnds
-        wrapUpEnds
-        highLevelNote
         phasedIn
         phasedInNote
         createdDts
