+++ src/tests/mock/readonly.ts
@@ -115,16 +115,6 @@ const modelBasicsData: GetAllBasicsTypes = {
   goal: 'To get more candy',
   testInterventions: 'The great candy machine',
   note: "The machine doesn't work yet",
-  completeICIP: '2022-06-03T19:32:24.412662Z',
-  clearanceStarts: '2022-06-03T19:32:24.412662Z',
-  clearanceEnds: '2022-06-03T19:32:24.412662Z',
-  announced: '2022-06-03T19:32:24.412662Z',
-  applicationsStart: '2022-06-03T19:32:24.412662Z',
-  applicationsEnd: '2022-06-03T19:32:24.412662Z',
-  performancePeriodStarts: '2022-06-03T19:32:24.412662Z',
-  performancePeriodEnds: '2022-06-03T19:32:24.412662Z',
-  wrapUpEnds: '2022-06-03T19:32:24.412662Z',
-  highLevelNote: 'Theses are my best guess notes',
   phasedIn: false,
   phasedInNote: "This can't be phased in",
   modifiedDts: '2022-06-03T19:32:24.412662Z',
